# Reduce number of optimization passes we run