The previous set was based on trying to get the absolute
best results on microbenchmarks; go back to the small set.

We'll want to revisit this later, but for now we're not getting much benefit from these (especially with type speculation not firing), so turning them off is an easy win:
```
      django_template2.py             5.2s (2)             5.1s (2)  -2.3%
            pyxl_bench.py             3.4s (2)             3.3s (2)  -2.1%
sqlalchemy_imperative2.py             4.0s (2)             3.9s (2)  -1.2%
                  geomean                 4.1s                 4.0s  -1.9%
```